[Regression] Selections are incorrectly invalidated
https://codereview.chromium.org/570763003 removed the wrong
codepath when selecting: |clipToVisibleContent| was true in
most cases and we removed this codepath.

This change just re-introduce the incorrectly removed
codepath. Most baselines are just reverted to their original
ones prior to the change. The SVG changes are new and
corresponds to the selection box bleeding out of the box,
which is a better behavior.

BUG=417242,416186

Review URL: https://codereview.chromium.org/598113003

git-svn-id: svn://svn.chromium.org/blink/trunk@182697 bbb929c8-8fbe-4397-9dbb-9b2b20218538