Let the RenderFlowThread layer have an empty size.
The RenderFlowThread has a RenderLayer used as a helper for painting. The
layer got its bounds from the RenderFlowThread contents which is typically
a tall strip and not the size of the area that where the contents are
rendered. That resulted in unnecessarily expanding the height of the parent
RenderLayer backing. The MultiColumnSet already contributes the correct
sizes, thus we can return an empty bounding rectangle for the layer of the
RenderFlowThread.

R=mstensho@opera.com
BUG=418899

Review URL: https://codereview.chromium.org/612363002

git-svn-id: svn://svn.chromium.org/blink/trunk@182906 bbb929c8-8fbe-4397-9dbb-9b2b20218538