Fix pixelated HiDPI Mac widgets with impl-side painting
When drawing the native controls to a SkiaBitLocker, tell the bit locker
about the target device scale factor, so that it will record a higher (or
lower, if appropriate) resolution bitmap to play back on the painting
thread.

Also, specify a clip rect for progress bars, since that was missed in or
updated after the previous patch.

The ability to create a SkiaBitLocker without a clip rect will be removed
soon.

BUG=408557

Review URL: https://codereview.chromium.org/617003002

git-svn-id: svn://svn.chromium.org/blink/trunk@183006 bbb929c8-8fbe-4397-9dbb-9b2b20218538