Fix more disabled MSVC warnings, Source/wtf/ edition.
Mostly, this is about fixing "possible value truncation", usually when assigning
a larger type to a smaller one.

The PartitionAlloc changes (size_t -> uint16_t in many places) are because
ultimately those values are getting read/written to a variety of fields in
PartitionPage, PartitionBucket, etc. that are uint16_ts.  Another possible fix
would be to make these fields size_t.  I haven't done this, under the assumption
that increasing the size of these objects might be bad, but I'd be happy to do
so.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/521543002

git-svn-id: svn://svn.chromium.org/blink/trunk@183172 bbb929c8-8fbe-4397-9dbb-9b2b20218538