Implement HTMLMarqueeElement's animation in private scripts
This CL completely migrates HTMLMarqueeElement's implementation to Blink-in-JS. This enables us to remove RenderMarquee.cpp and most of HTMLMarqueeElement.cpp.

The only regression this migration may cause is a case where HTMLMarqueeElement's attributes are updated during an animation. If you update 'behavior', 'direction', 'loop', 'scrollAmount', 'scrollDelay' or 'trueSpeed' during the animation, the animation is expected to be continued with the new attribute setting. However, in this CL, the animation is reset with the new attribute setting. I'm not sure if this regression causes any real issue (I hope not). I think we can land this and see how it goes.

This CL slightly changes the rendering of the HTMLMarqueeElement. I updated the following layout tests accordingly.

[fast/block/float/marquee-shrink-to-avoid-floats.html]
The old rendering is wrong, and the new rendering is correct. The old rendering matches WebKit. The new rendering matches the spec, IE and Firefox.

[fast/css/MarqueeLayoutTest.html]
The old rendering is wrong, and the new rendering is correct. The old rendering matches WebKit and IE. The new rendering matches the spec and Firefox.

[html/marquee-scroll.html]
Blink-in-JS uses Web animation APIs and it renders an animation more smoothly than the previous C++ implementation. As a result, the animation moves more quickly than before and thus makes the image result undeterministic. Thus I changed the color of characters displayed in the marquee element equal to the background color of the marquee element, just to make the image result timing-independent.

[fast/inline-block/003.html]
Ditto.

[http/tests/activedomobject/marquee.html]
Now that the HTMLMarqueeElement is not an ActiveDOMObject. Thus I changed the number of ActiveDOMObjects in the test.

BUG=341031

Review URL: https://codereview.chromium.org/394773003

git-svn-id: svn://svn.chromium.org/blink/trunk@183245 bbb929c8-8fbe-4397-9dbb-9b2b20218538