Mixed Content: Use a static method to check form actions.
In order to break MixedContentChecker's dependency on LocalFrame, we
need to replace the existing checking mechanisms with static calls. This
patch moves the logic from the existing ::canSubmitToInsecureForm method
into a new, static ::checkFormAction method.

It also drops a test that doesn't represent functionality we want to have
at the moment: we don't block form submission based on this, and we don't
plan to until we have some data about what impact that would have on the
web. Until that time, we don't need a test for that functionality.

Also also, this corrects the FormsSubmitted counter. That stat ought to
count all form submissions as "FormsSubmitted", and mixed submissions as
a subset thereof (e.g. 5 form submissions, with 2 of them mixed would
result in FormsSubmitted: 5, MixedContentFormsSubmitted: 2. Currently, it
results in 3 and 2, respectively).

------------------------------------------------------------------------
This was originally landed in https://codereview.chromium.org/566533003,
but reverted due to hard-to-track-down ChromeOS issues. I'll re-land it
in parts to narrow down the bits that caused problems.
------------------------------------------------------------------------

BUG=411184
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/620063002

git-svn-id: svn://svn.chromium.org/blink/trunk@183301 bbb929c8-8fbe-4397-9dbb-9b2b20218538