Correct several virtual/gpu expectations.
With revision 183577 several virtual gpu pixel expectations were
updated with obviously incorrect images. This appears to be due to
some sort of race in Release which needs to be investigated.

The new baselines in this change are obviously required, and the Debug
bot is correctly failing, but the Release bot appears to believe the
old baslines are ok, which they are not.

git-svn-id: svn://svn.chromium.org/blink/trunk@183579 bbb929c8-8fbe-4397-9dbb-9b2b20218538