bindings: Allows the test IDLs to override the production IDLs.
Some of the interface names are hard-coded in bindings/templates/*.{h,cpp}.
For example, Window, ArrayBuffer, etc.  So it will be convenient if we could
have the same interface name in the production IDLs and test IDLs.

This CL allows us to have the same interface name in both of the production
and test IDLs, and in that case, the test IDLs override the production IDLs
when testing.

Review URL: https://codereview.chromium.org/638143006

git-svn-id: svn://svn.chromium.org/blink/trunk@184059 bbb929c8-8fbe-4397-9dbb-9b2b20218538