Correcting -expected.png files generated after rebaseline.
These tests were initialy rebaselined as a part of,
https://codereview.chromium.org/568453002/

The -expected.png files generated for Win-XP,
as a part of rebaseline are upside down.
Please refer here,
https://codereview.chromium.org/662683004/

The original change should not cause this flanky behavior
of tests. It may be due to flanky behavior of build bot.
So requesting rebaseline for these tests again.

BUG=425157

Review URL: https://codereview.chromium.org/663033003

git-svn-id: svn://svn.chromium.org/blink/trunk@184102 bbb929c8-8fbe-4397-9dbb-9b2b20218538