Fix missing text update of RenderMenuList with external popup menu
The issue is caused by that:
- ExternalPopupMenu::didAcceptIndices() calls
  PopupMenuClient::listBoxSelectItem() before closing the popup;
and
- RenderMenuList::updateFromElement() doesn't update the button text
  if popup is visible, and doesn't update button text when the popup
  is being closed.

Combination of the above causes missing text update when the external
popup menu is being closed. (The issue is sometimes hidden if there
is pending styleRecalc/layout when the popup menu is closed because
handling of the styleRecalc/layout will update the text.)

Fixing either of the above will fix the bug. To ensure consistency
of ExtenalPopupMenu::didAcceptIndices() and
ExternalPopupMenu::didAcceptIndex(), and ensure status correctness of
RenderMenuList, both are fixed in the patch.

BUG=384600
TEST=ExternalPopupMenuTest.*

Review URL: https://codereview.chromium.org/668843002

git-svn-id: svn://svn.chromium.org/blink/trunk@184106 bbb929c8-8fbe-4397-9dbb-9b2b20218538