Load high-res viewport-dependent images on resize for all source-sizes.
Earlier CLs that added the loading of a higher resolution resource when
viewport size varies, did that only when the source-size is viewport
dependent itself, with the rational that this is the main use-case for
this behavior. That was a mistake, and users have been complaining.

This CL simply enables the resize notifications regardless of the viewport
dependency of the source-size value.

BUG=424178

Review URL: https://codereview.chromium.org/657213003

git-svn-id: svn://svn.chromium.org/blink/trunk@184112 bbb929c8-8fbe-4397-9dbb-9b2b20218538