Implement layout invalidation tracking in devtools
This patch adds layout invalidation tracking to the invalidation
tracking experiment. Our full design doc is at:
https://docs.google.com/document/d/1BXT3_gD--YdbIYMlTzNGsyUUMBCZJ7V5qGJhMA-pGrs/edit

This patch implements the invalidation tracking logic for layout
events. All layout invalidations are recorded and when a layout
occurs they are collected and displayed to the user. Style invalidations
occurring at the same time are not shown.

This patch is a fairly straightforward extension of [1] but for
layout events. The InspectorTest.findTimelineRecord function has been
added for capturing the N-th event of a specific type and all other
inspector tests updated to use it.

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=184212

BUG=410701

Review URL: https://codereview.chromium.org/671913005

git-svn-id: svn://svn.chromium.org/blink/trunk@184269 bbb929c8-8fbe-4397-9dbb-9b2b20218538