CSS Tokenizer: Fix handling of escaped newlines
This patch fixes our handling of escaped newlines in the css-syntax
tokenizer. Since we don't perform preprocessing, the logic is slightly
trickier than in the spec. The preprocessing replaces \r, \r\n, \f with
\n.

Regarding escaped newlines in strings, the spec states:
"Otherwise, if the next input code point is a newline, consume it."
In this case, we may need to consume two tokens if we have \r\n.

The other cases don't require special handling, since \r\n starts with
\r (also a newline) and in these cases the following token is just going
to be a whitespace token.

BUG=424988

Review URL: https://codereview.chromium.org/656033010

git-svn-id: svn://svn.chromium.org/blink/trunk@184347 bbb929c8-8fbe-4397-9dbb-9b2b20218538