Service Worker: Remove invalid test case for Cache.put() rejection
The test case was asserting that 'get' was not a valid method for
Cache.put(). But per the fetch spec, the Request's method is
normalized during construction so 'get' becomes 'GET' and that's
valid. There's a later test case which covers this using 'HEAD', so
just remove it.

BUG=374802
R=asanka@chromium.org

Review URL: https://codereview.chromium.org/678513002

git-svn-id: svn://svn.chromium.org/blink/trunk@184372 bbb929c8-8fbe-4397-9dbb-9b2b20218538