Set keyIdentifier to "Enter" instead of "U+000A".
When Ctrl+Enter is pressed, the keypress event is triggered with keycode 0x0A
instead of 0x0D (VKEY_RETURN). Consequently the keyIdentifier becomes "U+000A"
instead of "Enter" in WebKeyboardEvent::setKeyIdentifierFromWindowsKeyCode().

This patch adds a mapping from 0x0A to "Enter". The original specification said
"If no appropriate mapping was found, create a key identifier as human friendly
as possible. As an example, the Enter key is mapped to the key identifier
"Enter" and not to "U+000A" (Line Feed), given that this key generates the
character 00000A on Unix operating systems and the characters 00000D and 00000A
on Windows operating systems."
http://www.w3.org/TR/2003/NOTE-DOM-Level-3-Events-20031107/keyset.html#Guide

BUG=425859

Review URL: https://codereview.chromium.org/672923002

git-svn-id: svn://svn.chromium.org/blink/trunk@184409 bbb929c8-8fbe-4397-9dbb-9b2b20218538