Fixes look up of the best clickable node.
There is a case that EventHandler::bestClickableNodeForHitTestResult
and Internals::touchNodeAdjustedToBestClickableNode return an
uninitialized pointer.  In that case, we must return null instead.

This CL fixes the issue and returns null of type Node* as expected.

Zero-initializations in Internals.cpp are not mandatory, but I think
it's better to have them just in case and as a good coding practice
(never leave pointers uninitialized).

Note that I cannot reproduce the issue 426760, so cannot confirm the
fix, too.  But I think this is the cause of the issue.

BUG=426760

Review URL: https://codereview.chromium.org/675403003

git-svn-id: svn://svn.chromium.org/blink/trunk@184439 bbb929c8-8fbe-4397-9dbb-9b2b20218538