Add live region root to the cached properties of an AXObject.
This adds the live region root to the set of properties of an
AXObject that are cached. The live region root is simply the
root object of the live region an object is contained in, if any.
Also exposes new "container" functions exposing the value of
live region properties of the live region root, for any given
node.

This will be followed by a Chromium change to use these new APIs.

BUG=427402

Review URL: https://codereview.chromium.org/659563006

git-svn-id: svn://svn.chromium.org/blink/trunk@184471 bbb929c8-8fbe-4397-9dbb-9b2b20218538